pyproject.toml
[project]
name = "revolut-gateway"
version = "0.1.0"
description = "Revolut Gateway - WebSocket LLM Gateway"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "openai>=1.60.0",
    "aiofiles>=24.1.0",
]

[project.scripts]
revolut-gateway = "main:run"
